requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
flatbuffers==23.1.21
gast==0.4.0
google-auth-oauthlib==0.4.6
google-auth==2.16.2
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
pip==23.0.1
protobuf==3.19.6
pyasn1-modules==0.2.8
pyasn1==0.4.8
requests-oauthlib==1.3.1
requests==2.28.2
rsa==4.9
setuptools==67.4.0
six==1.16.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.11.2
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow==2.11.0
termcolor==2.2.0
typing-extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wheel==0.38.4
wrapt==1.15.0
zipp==3.15.0
matplotlib==3.7.1